% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert}
\alias{assert}
\title{Simple error assertions}
\usage{
assert(expr, error)
}
\arguments{
\item{expr}{logical expression}

\item{error}{character, error message}
}
\value{
An error message, or nothing
}
\description{
Simple error assertions
}
