% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{assign_baseline}
\alias{assign_baseline}
\title{Append baseline data to initial data}
\usage{
assign_baseline(baseline = NULL, data)
}
\arguments{
\item{baseline}{character value of the default baseline to assign}

\item{data}{a data.frame to append the baseline to}
}
\value{
the data.frame passed by the user with an additional variable appended
}
\description{
Append baseline data to initial data
}
