% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{assign_grade}
\alias{assign_grade}
\title{Assign student a grade}
\usage{
assign_grade(age, ability)
}
\arguments{
\item{age}{age of the student in years}

\item{ability}{a modifier that signifies student ability?}
}
\value{
a vector of grade levels
}
\description{
Assign student a grade
}
\examples{
age <- c(9, 10, 11, 12)
assign_grade(age = age)
}
