% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{better_sim.lm}
\alias{better_sim.lm}
\title{Generate simulated predictions from a linear model that account for
model and parameter error}
\usage{
better_sim.lm(object, nsim, newdata, resid_error = FALSE)
}
\arguments{
\item{object}{an object of class lm or glm}

\item{nsim}{number of simulations per observation to generate}

\item{newdata}{dataframe containing the observations to generate predictions
for}

\item{resid_error}{should the model residual error be added to predictions,
default is FALSE}
}
\value{
a matrix of predictions nrow(newdata) x nsim columns
}
\description{
Generate simulated predictions from a linear model that account for
model and parameter error
}
