% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{ceds_cleaner}
\alias{ceds_cleaner}
\title{CEDS oputput function}
\usage{
ceds_cleaner(simouts, output = "directory", directory = NULL)
}
\arguments{
\item{simouts}{a simulation list resulting from a call to the \code{\link{simpop}}
function}

\item{output}{a character, default "directory", specifying where the output should good}

\item{directory}{a path to a directory to store the output files}
}
\value{
Nothing. Output is saved out to a file on disk.
}
\description{
CEDS oputput function
}
