% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{cond_prob}
\alias{cond_prob}
\title{Generate conditional probabilities by group}
\usage{
cond_prob(data, factor, newvar, prob_list)
}
\arguments{
\item{data}{dataframe to add variable to}

\item{factor}{grouping variable that probability of \code{newvar} is conditional on}

\item{newvar}{name, character, of new variable defined by \code{prob_list}}

\item{prob_list}{a list, defining the way \code{newvar} should be generated}
}
\value{
data.frame with \code{newvar} appended to dataframe
}
\description{
Generate conditional probabilities by group
}
