% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_grade}
\alias{convert_grade}
\title{Convert auto-generated grades to better grades}
\usage{
convert_grade(x)
}
\arguments{
\item{x}{a vector of grades formatted with g}
}
\value{
a character vector of grades
}
\description{
Convert auto-generated grades to better grades
}
