% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_transitions.R
\name{createAutocorBinSeries}
\alias{createAutocorBinSeries}
\title{Create an autocorrelated binary series}
\usage{
createAutocorBinSeries(n = 100, mean = 0.5, corr = 0)
}
\arguments{
\item{n}{integer, length of series to generate}

\item{mean}{numeric between 0 and 1, proportion of cases with value 1}

\item{corr}{numeric between -1 and 1, how correlated should series be?}
}
\value{
A binary series
}
\description{
Generate a binary series that is autocorrelated using the 
Markov method from \code{\link{createSeries}}
}
\examples{
createAutocorBinSeries(n=12,mean=0.5,corr=0.9)
createAutocorBinSeries(n=100,mean=0.5,corr=0.1)
}
