% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{diag_offset}
\alias{diag_offset}
\title{Offset the diagonal values of a matrix}
\usage{
diag_offset(matrix, offset = 1L)
}
\arguments{
\item{matrix}{a square matrix}

\item{offset}{an integer value to offset the diagonal by}
}
\value{
a square matrix
}
\description{
Offset the diagonal values of a matrix
}
