% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{expand.grid.df}
\alias{expand.grid.df}
\title{Expand dataframe into a complete grid}
\usage{
expand.grid.df(...)
}
\arguments{
\item{...}{dataframe}
}
\value{
an expanded data frame
}
\description{
Expand dataframe into a complete grid
}
\details{
From SO {\link{http://stackoverflow.com/questions/11693599/alternative-to-expand-grid-for-data-frames}}
}
