% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_grid_df}
\alias{expand_grid_df}
\title{Expand dataframe into a complete grid}
\usage{
expand_grid_df(...)
}
\arguments{
\item{...}{dataframe}
}
\value{
an expanded data frame
}
\description{
Expand dataframe into a complete grid
}
\details{
From SO {\url{http://stackoverflow.com/questions/11693599/alternative-to-expand-grid-for-data-frames}}
}
