% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_transitions.R
\name{findTransitions}
\alias{findTransitions}
\title{Identify the parameters that define a series of binary outcomes}
\usage{
findTransitions(series, return = c("matrix", "simple"))
}
\arguments{
\item{series}{a vector of 0 and 1 values}

\item{return}{a character with two options, matrix returns a transition 
matrix, simple returns a list with two elements, mean and cor}
}
\value{
Either a transition matrix or a list with parameters mean and cor 
defining the transitions in the vector
}
\description{
Identify the parameters that define a series of binary outcomes
}
\examples{
series <- createSeries(10, matrix(c(0.444, 0.111, 0.222, 0.222), 
                       nrow = 2, byrow =TRUE))
createSeries(10, findTransitions(series))
}
