% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_transitions.R
\name{fit_series}
\alias{fit_series}
\title{Identify the parameters that define a series of binary outcomes}
\usage{
fit_series(series, return = c("matrix", "fit"), ...)
}
\arguments{
\item{series}{a vector of 0 and 1 values}

\item{return}{a character with two options, matrix returns a transition
matrix, "fit" returns a \code{\link{markovchain}} object}

\item{...}{additional arguments to pass to \code{\link{markovchainFit}}}
}
\value{
Either a transition matrix or a list with parameters mean and cor
defining the transitions in the vector
}
\description{
Identify the parameters that define a series of binary outcomes
}
\examples{
series <- make_markov_series(10, matrix(c(0.3, 0.7, 0.25, 0.75),
                       nrow = 2, byrow =TRUE))
fit_series(series)
}
