% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_annual_gpa}
\alias{gen_annual_gpa}
\title{Generate annual GPA sequence for students based on final GPA}
\usage{
gen_annual_gpa(gpa_ontrack)
}
\arguments{
\item{gpa_ontrack}{a data.frame containing final GPA for a student}
}
\value{
\code{gpa_ontrack} appended with gpa by year
}
\description{
Generate annual GPA sequence for students based on final GPA
}
