% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{gen_annual_status}
\alias{gen_annual_status}
\title{Generate annual status trajectories per student}
\usage{
gen_annual_status(data, control = sim_control())
}
\arguments{
\item{data}{student-year data}

\item{control}{control list}
}
\value{
the \code{data} object, with additional variables appended
}
\description{
Generate annual status trajectories per student
}
