% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_assess}
\alias{gen_assess}
\title{Generate an assessment table}
\usage{
gen_assess(data, control = sim_control())
}
\arguments{
\item{data}{student-year data}

\item{control}{output from \code{sim_control}}
}
\value{
a two-column dataframe with math and reading scores
}
\description{
Generate an assessment table
}
