% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_credits}
\alias{gen_credits}
\title{Generate credit sequence for students based on their final GPA}
\usage{
gen_credits(gpa_ontrack)
}
\arguments{
\item{gpa_ontrack}{a data.frame containing final GPA for a student}
}
\value{
\code{gpa_ontrack} but appended with credits by year
}
\description{
Generate credit sequence for students based on their final GPA
}
