% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_gpa}
\alias{gen_gpa}
\title{Generate a final GPA for students}
\usage{
gen_gpa(data, control = sim_control())
}
\arguments{
\item{data}{a dataframe with variables}

\item{control}{a sim_control parmeter, default is \code{sim_control}}
}
\value{
a numeric vector
}
\description{
Generate a final GPA for students
}
