% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_grad}
\alias{gen_grad}
\title{Generate a high school graduation for students}
\usage{
gen_grad(data, control = sim_control())
}
\arguments{
\item{data}{a dataframe with variables}

\item{control}{a sim_control parmeter, default is \code{sim_control}}
}
\value{
a data.frame with two values, a probability and a binary outcome
}
\description{
Generate a high school graduation for students
}
