% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_hs_annual}
\alias{gen_hs_annual}
\title{Generate annual HS outcomes}
\usage{
gen_hs_annual(hs_outcomes, stu_year)
}
\arguments{
\item{hs_outcomes}{a data frame with final gpa, hs_status, and sid}

\item{stu_year}{a data frame with student enrollment}
}
\value{
an expanded table with credits and gpa information
}
\description{
Generate annual HS outcomes
}
