% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{gen_initial_status}
\alias{gen_initial_status}
\title{Generate initial student status indicators}
\usage{
gen_initial_status(data, baseline)
}
\arguments{
\item{data}{that includes the pre-requsites for generating each status}

\item{baseline}{character, name of a baseline status to calculate}
}
\value{
the data with status variables appended
}
\description{
Generate initial student status indicators
}
