% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_ontrack}
\alias{gen_ontrack}
\title{Calculate on_track based on credits}
\usage{
gen_ontrack(gpa_ontrack)
}
\arguments{
\item{gpa_ontrack}{a data.frame with cumulative credits}
}
\value{
\code{gpa_ontrack} with ontrack indicators appended
}
\description{
Calculate on_track based on credits
}
