% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{gen_outcome_model}
\alias{gen_outcome_model}
\title{Generate data for a multilevel outcome}
\usage{
gen_outcome_model(fixed, fixed_param, random_var, fact_vars, cov_param = NULL,
  cor_vars = NULL, ngrps, unbalanceRange)
}
\arguments{
\item{fixed}{a formula with RHS only that specifies the variables to use}

\item{fixed_param}{a vector of numerics for the coefficients of variables in fixed}

\item{random_var}{a numeric, length 1, variance of random (school level) component}

\item{fact_vars}{for each variable in fixed that is a factor, a definition...}

\item{cov_param}{a list, defining any continuous variables}

\item{cor_vars}{correlations between fixed variables}

\item{ngrps}{number of schools}

\item{unbalanceRange}{range of enrollments in each school}
}
\value{
a data.frame
}
\description{
Generate data for a multilevel outcome
}
\examples{
zed2 <- gen_outcome_model(fixed = ~ 1 + gifted.f + iep.f + frpl.f + ell.f + male.f,
random_var = 0.77, fixed_param = c(1.06, 0.72, -.2, -0.513, -0.4559, -0.356),
fact_vars = list(numlevels = c(2, 2, 2, 2, 2), var_type = c(rep('lvl1', 5))),
ngrps = 20, unbalanceRange = c(75, 900))
}
