% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_ps}
\alias{gen_ps}
\title{Generate postsecondary enrollment outcome}
\usage{
gen_ps(data, control = sim_control())
}
\arguments{
\item{data}{cohort data}

\item{control}{output from \code{sim_control}}
}
\value{
a two-column dataframe with probabilities and binary outcome
}
\description{
Generate postsecondary enrollment outcome
}
