% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_ps_enrollment}
\alias{gen_ps_enrollment}
\title{Generate a student-year long table of postsecondary enrollments}
\usage{
gen_ps_enrollment(hs_outcomes, nsc)
}
\arguments{
\item{hs_outcomes}{a dataframe of high school outcomes}

\item{nsc}{a dataframe of postsecondary institutions}
}
\value{
a table of enrollments
}
\description{
Generate a student-year long table of postsecondary enrollments
}
