% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{gen_schools}
\alias{gen_schools}
\title{Generate a roster of schools to assign students to}
\usage{
gen_schools(n, mean = NULL, sigma = NULL, names = NULL)
}
\arguments{
\item{n}{number of schools}

\item{mean}{a vector of means for the school attributes}

\item{sigma}{a covariance matrix for the school attributes}

\item{names}{vector to draw names from}
}
\value{
a data.frame with schools and their attributes
}
\description{
Generate a roster of schools to assign students to
}
