% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{gen_student_years}
\alias{gen_student_years}
\title{Generate annual student observations}
\usage{
gen_student_years(data, control = sim_control())
}
\arguments{
\item{data}{students to generate annual data for}

\item{control}{a list, defined by \code{\link{sim_control}}}
}
\value{
a data.frame
}
\description{
Generate annual student observations
}
