% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{gen_students}
\alias{gen_students}
\title{Generate student-level attributes}
\usage{
gen_students(n, seed, race_groups = NULL, race_prob = NULL,
  ses_list = NULL)
}
\arguments{
\item{n}{integer, number of students to simulate}

\item{seed}{integer, random seed to use}

\item{race_groups}{vector of labels for race groups}

\item{race_prob}{vector of numerics, same length as \code{race_groups}}

\item{ses_list}{a probability list}
}
\value{
a data.frame
}
\description{
Generate student-level attributes
}
