% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{get_baseline}
\alias{get_baseline}
\title{Function to load in SDP default baseline data}
\usage{
get_baseline(bl)
}
\arguments{
\item{bl}{a character naming the type of baseline available}
}
\value{
the restored baseline object
}
\description{
Function to load in SDP default baseline data
}
