% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDS.R
\name{get_code_values}
\alias{get_code_values}
\title{Get values of codes from the CEDS Crosswalk List}
\usage{
get_code_values(x)
}
\arguments{
\item{x}{the data.frame character element that contains the codes}
}
\value{
a list with the labels and levels properly formatted
}
\description{
Get values of codes from the CEDS Crosswalk List
}
