% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{get_sim_groupvars}
\alias{get_sim_groupvars}
\title{Get grouping variables}
\usage{
get_sim_groupvars(control = sim_control())
}
\arguments{
\item{control}{a control list produced by \code{\link{sim_control}}}
}
\value{
a character vector of grouping terms in control lists
}
\description{
Get grouping variables
}
