% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{grade_transitions}
\alias{grade_transitions}
\title{Generate a grade advancement transition matrix}
\usage{
grade_transitions(ngrades = 15L, diag_limit = 0.975)
}
\arguments{
\item{ngrades}{integer, the number of grade levels to simulate}

\item{diag_limit}{the minimum probability of a student advancing to the next grade}
}
\value{
a matrix, ngrades x ngrades of grade transition probabilities
}
\description{
Generate a grade advancement transition matrix
}
