% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_rescale}
\alias{group_rescale}
\title{Rescaling variables in groups}
\usage{
group_rescale(data, var, group_var, newvar = NULL)
}
\arguments{
\item{data}{a dataframe containing variables you wish to rescale}

\item{var}{name of the variable to be rescaled}

\item{group_var}{character vector of the grouping terms}

\item{newvar}{optional character vector for the name of the new rescaled variable}
}
\value{
data with the newvar appended
}
\description{
Rescaling variables in groups
}
