% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m_sum}
\alias{m_sum}
\title{Sum matrix elements in a list}
\usage{
m_sum(l)
}
\arguments{
\item{l}{a list}
}
\value{
a list of summed matrix elements
}
\description{
Sum matrix elements in a list
}
