% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDS.R
\name{make_inds}
\alias{make_inds}
\title{Append indicator variables to a data frame based on a single factor variable}
\source{
\url{http://stackoverflow.com/questions/35943455/creating-indicator-variable-columns-in-dplyr-chain}
}
\usage{
make_inds(data, col)
}
\arguments{
\item{data}{a data frame}

\item{col}{character, name of the factor column to generate indicators for}
}
\value{
a data frame with the factor levels appended as columns
}
\description{
Append indicator variables to a data frame based on a single factor variable
}
