% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_transitions.R
\name{make_markov_series}
\alias{make_markov_series}
\title{Create a series of binary variables based on a transition matrix}
\source{
\itemize{
\item \url{http://stats.stackexchange.com/questions/14175/how-to-generate-random-auto-correlated-binary-time-series-data}
\item \url{https://en.wikipedia.org/wiki/Examples_of_Markov_chains}
}
}
\usage{
make_markov_series(n, tm, burnin = NULL, ...)
}
\arguments{
\item{n}{an integer representing the length of the series}

\item{tm}{a Transition Matrix that describes the transition matrix}

\item{burnin}{integer, number of values to simulate before drawing from the series}

\item{...}{additional arguments to apss to \code{\link{markovchainSequence}}}
}
\value{
a vector of length \code{n} with a random series of 0 and 1 generated
from the \code{tm}
}
\description{
Generate a
}
\details{
A Transition Matrix is a 2x2 matrix where:
\itemize{
 \item Element 1, 1 is the probability of moving to state 0 conditional on being in state 0
 \item Element 2, 1 is the probability of moving to state 0 conditional on being in state 1
 \item Element 1, 2 is the probability of moving to state 1 conditional on being in state 0
 \item Element 2, 2 is the probability of moving to state 1 conditional on being in state 1
}
}
\examples{
make_markov_series(10, matrix(c(0.6,0.4,0.9,0.1), nrow=2, byrow=TRUE))
}
