% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{map_CEDS}
\alias{map_CEDS}
\title{Map to and from CEDS column names}
\usage{
map_CEDS(user, category = NULL, CEDS = NULL)
}
\arguments{
\item{user}{character vector of variable names to match}

\item{category}{character, a category of CEDS data to match}

\item{CEDS}{optional}
}
\value{
mapped CEDS names
}
\description{
Map to and from CEDS column names
}
