% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_control.R
\name{namedList}
\alias{namedList}
\title{Create a named list}
\usage{
namedList(...)
}
\arguments{
\item{...}{arguments to pass to the list}
}
\value{
a named list
}
\description{
Create a named list
}
