% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_clip}
\alias{num_clip}
\title{Clip a vector to be between a minimum and a maximum}
\usage{
num_clip(x, min, max)
}
\arguments{
\item{x}{a numeric vector}

\item{min}{numeric, a floor}

\item{max}{numeric, a ceiling}
}
\value{
x, truncated to be between min and max
}
\description{
Clip a vector to be between a minimum and a maximum
}
