% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_grade}
\alias{num_grade}
\title{Convert a character representation of grades to numeric}
\usage{
num_grade(grade)
}
\arguments{
\item{grade}{a character vector of grades with character labels}
}
\value{
a numeric vector, length of grade, representing grade levels as numbers
}
\description{
Convert a character representation of grades to numeric
}
