% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{popsim_control}
\alias{popsim_control}
\title{Grand simulation}
\usage{
popsim_control(n, seed, race_groups = NULL, race_prob = NULL,
  ses_list = NULL, minyear = 1997, maxyear = 2017)
}
\arguments{
\item{n}{integer, number of students to simulate}

\item{seed}{integer, random seed to use}

\item{race_groups}{vector of labels for race groups}

\item{race_prob}{vector of numerics, same length as \code{race_groups}}

\item{ses_list}{a probability list}

\item{minyear}{integer, year to start data}

\item{maxyear}{integer, year to end data}
}
\description{
Grand simulation
}
\examples{
out <- popsim_control(20, seed = 213)
}
