% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rand_vect_cont}
\alias{rand_vect_cont}
\title{Generate positive random numbers that sum to a value}
\usage{
rand_vect_cont(N, M, sd = 1)
}
\arguments{
\item{N}{number of numbers to generate}

\item{M}{constraint on the sum}

\item{sd}{standard deviation of values}
}
\value{
a vector of numerics, length N
}
\description{
Generate positive random numbers that sum to a value
}
\examples{
out <- rand_vect_cont(N = 10, M = 2, sd = 1)
sum(out)
length(out)
}
