% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_credits}
\alias{recode_credits}
\title{Recode numeric data into plausible credit data}
\usage{
recode_credits(x, top = 6)
}
\arguments{
\item{x}{numeric, vector of numerics to be truncated}

\item{top}{integer, maximum allowable value}
}
\value{
a vector, length of x, with truncated values
}
\description{
Recode numeric data into plausible credit data
}
\details{
enforces no negative numbers and truncates data at a user specified
maximum
}
