% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDS.R
\name{recode_options}
\alias{recode_options}
\title{Recode options}
\usage{
recode_options(data, from = c("SDP", "CEDS"))
}
\arguments{
\item{data}{a data.frame to recode the variables to CEDS from}

\item{from}{the data definitions you are recoding from}
}
\value{
the data object, but with all values matching CEDS specification
recoded to meet the CEDS specification
}
\description{
Recode options
}
