% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Reliabily rescale numerics with missingness}
\usage{
rescale(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
x mean centered and divided by it's standard deviation
}
\description{
Reliabily rescale numerics with missingness
}
