% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{rescale_gpa}
\alias{rescale_gpa}
\title{Rescale scaled GPA to be on 0-4 scale}
\usage{
rescale_gpa(x)
}
\arguments{
\item{x}{a scaled normal variable}
}
\value{
a GPA rounded to tenths, from 0 to 1
}
\description{
Rescale scaled GPA to be on 0-4 scale
}
