% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{school_transitions}
\alias{school_transitions}
\title{Generate a random transition matrix for school enrollments}
\usage{
school_transitions(nschls = 15L, diag_limit = 0.9)
}
\arguments{
\item{nschls}{integer, the number of schools available, default = 15}

\item{diag_limit}{a numeric between 0 and 1 that sets the minimum probability that a student
will stay schools}
}
\value{
a transition matrix nschls X nschls of transition probabilities
}
\description{
Generate a random transition matrix for school enrollments
}
\examples{
out <- school_transitions(12)
out
}
