% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{sdp_cleaner}
\alias{sdp_cleaner}
\title{Convert a population simulation into an analysis file}
\usage{
sdp_cleaner(simouts)
}
\arguments{
\item{simouts}{result of the \code{simpop} function}
}
\value{
an analysis file as a single dataframe for HS outcomes
}
\description{
Convert a population simulation into an analysis file
}
