% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{sim_control}
\alias{sim_control}
\title{Set control parameters for simulated data}
\usage{
sim_control(race_groups = NULL, race_prob = NULL, ses_list = NULL,
  minyear = 1997, maxyear = 2017, n_cohorts = NULL)
}
\arguments{
\item{race_groups}{vector of labels for race groups}

\item{race_prob}{vector of numerics, same length as \code{race_groups}}

\item{ses_list}{a probability list}

\item{minyear}{an integer}

\item{maxyear}{an integer}
}
\value{
a named list
}
\description{
Set control parameters for simulated data
}
