% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tm_convert}
\alias{tm_convert}
\title{Covnert a matrix to a row-wise transition matrix}
\usage{
tm_convert(matrix)
}
\arguments{
\item{matrix}{a matrix tof counts}
}
\value{
matrix M divided by the sum of its rows
}
\description{
Covnert a matrix to a row-wise transition matrix
}
\examples{
base_mat <- structure(c(44985, 740, 781, 7640), .Dim = c(2L, 2L),
.Dimnames = list(c("0", "1"), c("0", "1")))
tm_convert(base_mat)
}
