% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unscale}
\alias{unscale}
\title{Unscale a scaled variable}
\usage{
unscale(x, mean, sd)
}
\arguments{
\item{x}{numeric vector that has been scaled}

\item{mean}{a numeric, the mean to add to x}

\item{sd}{a numeric, the standardized factor to divide x by}
}
\value{
x rescaled with mean and sd specified by the user
}
\description{
Unscale a scaled variable
}
