% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_probability_list}
\alias{validate_probability_list}
\title{Validate probability list formatting and structure}
\usage{
validate_probability_list(list)
}
\arguments{
\item{list}{list that is to be passed as a probability list to simulation functions}
}
\value{
Logical TRUE if list is valid, error if list is not valid.
}
\description{
Validate probability list formatting and structure
}
