% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_sim_parameter}
\alias{validate_sim_parameter}
\title{Validate outcome simulation list formatting and structure}
\usage{
validate_sim_parameter(list)
}
\arguments{
\item{list}{list that is to be passed as a parameter list to outcome simulation functions}
}
\value{
Logical TRUE if list is valid, error if list is not valid.
}
\description{
Validate outcome simulation list formatting and structure
}
