% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeroNA}
\alias{zeroNA}
\title{Convert NA values to 0 in a vector}
\usage{
zeroNA(x)
}
\arguments{
\item{x}{a numeric vector containing NAs}
}
\value{
a numeric vector where all NA values are 0
}
\description{
Convert NA values to 0 in a vector
}
