% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{assign_hs_outcomes}
\alias{assign_hs_outcomes}
\title{Assign high school outcomes}
\usage{
assign_hs_outcomes(g12_cohort, control = sim_control())
}
\arguments{
\item{g12_cohort}{a dataframe with certain high school attributes}

\item{control}{control parameters from the \code{sim_control()} function}
}
\value{
an outcome dataframe
}
\description{
Assign high school outcomes
}
