% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_outcomes.R
\name{gen_outcome_model}
\alias{gen_outcome_model}
\title{Generate data for a multilevel outcome}
\usage{
gen_outcome_model(fixed, fixed_param, random_var, fact_vars, cov_param = NULL,
  cor_vars = NULL, ngrps, unbalanceRange, type = "binary",
  with_err_gen = NULL, error_var = NULL)
}
\arguments{
\item{fixed}{a formula with RHS only that specifies the variables to use}

\item{fixed_param}{a vector of numerics for the coefficients of variables in fixed}

\item{random_var}{a numeric, length 1, variance of random (school level) component}

\item{fact_vars}{for each variable in fixed that is a factor, a definition...}

\item{cov_param}{a list, defining any continuous variables}

\item{cor_vars}{correlations between fixed variables}

\item{ngrps}{number of schools}

\item{unbalanceRange}{range of enrollments in each school}

\item{type}{character, either "binary" or "linear" to choose outcome variable
type to generate}

\item{with_err_gen}{name of a distribution function to generate the errors, optional}

\item{error_var}{integer values to pass to err_gen, optional}
}
\value{
a list with two elements
}
\description{
Generate data for a multilevel outcome
}
\examples{
zed2 <- do.call(gen_outcome_model, sim_control()$gpa_sim_parameters)
}
