% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{markov_cond_list}
\alias{markov_cond_list}
\title{Compute conditional Markov sequences in a pipeline}
\usage{
markov_cond_list(groupname, n, lst, ...)
}
\arguments{
\item{groupname}{variable that identifies the list element to use}

\item{n}{number of elements to generate, usually defined by \link[dplyr]{n} in dplyr}

\item{lst}{probability list that contains function and function parameters}

\item{...}{additional arguments passed into the function}
}
\value{
A sequence generated by functions in the probability list
}
\description{
Compute conditional Markov sequences in a pipeline
}
