% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{sim_control}
\alias{sim_control}
\title{Set control parameters for simulated data}
\usage{
sim_control(nschls = 2L, race_groups = NULL, race_prob = NULL,
  ses_list = NULL, minyear = 2002, maxyear = 2017, n_cohorts = NULL,
  gifted_list = NULL, iep_list = NULL, ell_list = NULL,
  grade_levels = NULL, school_means = NULL, school_cov_mat = NULL,
  school_names = NULL, gpa_sim_parameters = NULL,
  grad_sim_parameters = NULL)
}
\arguments{
\item{nschls}{integer for number of schools to create, default is 2}

\item{race_groups}{vector of labels for race groups}

\item{race_prob}{vector of numerics, same length as \code{race_groups}}

\item{ses_list}{a probability list}

\item{minyear}{an integer}

\item{maxyear}{an integer}

\item{n_cohorts}{number of cohorts to produce}

\item{gifted_list}{a probability list}

\item{iep_list}{a probability list}

\item{ell_list}{a probability list}

\item{grade_levels}{a probability list}

\item{school_means}{a named vector of means for school level attributes}

\item{school_cov_mat}{a covariance matrix for the school level attributes}

\item{school_names}{a vector to draw school names from}

\item{gpa_sim_parameters}{a list of parameters to pass to \code{gen_outcome_model}}

\item{grad_sim_parameters}{a list of parameters to pass to \code{gen_outcome_model}}
}
\value{
a named list
}
\description{
Set control parameters for simulated data
}
