% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_defaults.R
\name{simpop}
\alias{simpop}
\title{Grand simulation}
\usage{
simpop(nstu, seed = NULL, control = sim_control())
}
\arguments{
\item{nstu}{integer, number of students to simulate}

\item{seed}{integer, random seed to make simulation reproducible across
sessions, optional}

\item{control}{a list, defined by \code{\link{sim_control}}}
}
\value{
a list with simulated data
}
\description{
Grand simulation
}
\examples{
\dontrun{
out <- simpop(nstu = 20, seed = 213)
}
}
