% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_sequence}
\alias{tidy_sequence}
\title{Tidy a two-state markov sequence for output}
\usage{
tidy_sequence(seq, states)
}
\arguments{
\item{seq}{a vector of sequence elements, with only two states}

\item{states}{a vector of length two naming both possible states}
}
\value{
a data.frame
}
\description{
Tidy a two-state markov sequence for output
}
\examples{
tidy_sequence(seq = c("Yes", "No", "No", "No", "Yes", "Yes"),
             states = c("Yes", "No"))
}
